fix:完善边界条件

+++ sky-server/src/main/java/com/sky/mapper/DishMapper.java
@@ -13,5 +13,5 @@ public interface DishMapper {
      * @return
      */
     @Select("select count(*) from dish where category_id = #{categoryId}")
-    Integer countByCategoryId(Integer categoryId);
+    Integer countByCategoryId(Long categoryId);
 }
